+++ databasez/core/connection.py
@@ -38,7 +38,7 @@ class Connection:
         self._force_rollback = force_rollback
         self.connection_transaction: typing.Optional[Transaction] = None
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def __aenter__(self) -> Connection:
         async with self._connection_lock:
             self._connection_counter += 1
@@ -65,6 +65,7 @@ class Connection:
                 raise e
         return self
 
+    @multiloop_protector(False)
     async def __aexit__(
         self,
         exc_type: typing.Optional[typing.Type[BaseException]] = None,
@@ -89,7 +90,7 @@ class Connection:
     def _loop(self) -> typing.Any:
         return self._database._loop
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def fetch_all(
         self,
         query: typing.Union[ClauseElement, str],
@@ -99,7 +100,7 @@ class Connection:
         async with self._query_lock:
             return await self._connection.fetch_all(built_query)
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def fetch_one(
         self,
         query: typing.Union[ClauseElement, str],
@@ -110,7 +111,7 @@ class Connection:
         async with self._query_lock:
             return await self._connection.fetch_one(built_query, pos=pos)
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def fetch_val(
         self,
         query: typing.Union[ClauseElement, str],
@@ -122,7 +123,7 @@ class Connection:
         async with self._query_lock:
             return await self._connection.fetch_val(built_query, column, pos=pos)
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def execute(
         self,
         query: typing.Union[ClauseElement, str],
@@ -136,7 +137,7 @@ class Connection:
             async with self._query_lock:
                 return await self._connection.execute(query, values)
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def execute_many(
         self, query: typing.Union[ClauseElement, str], values: typing.Any = None
     ) -> typing.Union[typing.Sequence[interfaces.Record], int]:
@@ -148,7 +149,7 @@ class Connection:
             async with self._query_lock:
                 return await self._connection.execute_many(query, values)
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def iterate(
         self,
         query: typing.Union[ClauseElement, str],
@@ -160,7 +161,7 @@ class Connection:
             async for record in self._connection.iterate(built_query, batch_size):
                 yield record
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def batched_iterate(
         self,
         query: typing.Union[ClauseElement, str],
@@ -172,7 +173,7 @@ class Connection:
             async for records in self._connection.batched_iterate(built_query, batch_size):
                 yield records
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def run_sync(
         self,
         fn: typing.Callable[..., typing.Any],
@@ -182,15 +183,15 @@ class Connection:
         async with self._query_lock:
             return await self._connection.run_sync(fn, *args, **kwargs)
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def create_all(self, meta: MetaData, **kwargs: typing.Any) -> None:
         await self.run_sync(meta.create_all, **kwargs)
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def drop_all(self, meta: MetaData, **kwargs: typing.Any) -> None:
         await self.run_sync(meta.drop_all, **kwargs)
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     def transaction(self, *, force_rollback: bool = False, **kwargs: typing.Any) -> "Transaction":
         return Transaction(weakref.ref(self), force_rollback, **kwargs)
 
@@ -200,7 +201,7 @@ class Connection:
         """The first layer (sqlalchemy)."""
         return self._connection.async_connection
 
-    @multiloop_protector(True)
+    @multiloop_protector(False)
     async def get_raw_connection(self) -> typing.Any:
         """The real raw connection (driver)."""
         return await self.async_connection.get_raw_connection()