+++ docs/release-notes.md
@@ -1,6 +1,17 @@
 # Release Notes
 
 
+## 0.9.7
+
+### Added
+
+- It is now possible to use connect(), disconnect() instead of a async contextmanager in multi-loop calls (multithreading).
+
+### Fixed
+
+- Database calls are forwarded to subdatabase when possible. This unbreaks using not the returned database object.
+- `force_rollback` works also in multi-loop call (multithreading).
+
 ## 0.9.6
 
 ### Fixed